Remove calls to GetConfiguration
GetConfiguration was meant more for testing with the original code base and has been causing some confusion in the community. Redistributing the DLL is not recommended because we will make changes over time that might make the implementation incompatible while the COM interface will remain working (unless deprecation is absolutely necessary, but that would only affect the return code).